+++ G1/CollisionRectangle.cpp
@@ -2,18 +2,18 @@
 
 namespace G1 {
 
-	CollisionRectangle::CollisionRectangle(double length, double height, double worldCoordinateX, double worldCoordinateY, float rot)
+	CollisionRectangle::CollisionRectangle(float length, float height, float worldCoordinateX, float worldCoordinateY, float rot)
 	{
 		this->length = length;
 		this->height = height;
 		this->setTransform(Transform(worldCoordinateX, worldCoordinateY, rot));
 		this->collisionType = 'r';
-		this->farthestPoint = sqrt((length*length) / 4.0f + (height*height) / 4.0f);
+		this->farthestPoint = sqrtf((length*length) / 4.0f + (height*height) / 4.0f);
 		this->nearestPoint = (length < height ? length / 2.0f : height / 2.0f);
 	}
 
 	Vector2 CollisionRectangle::getFarthestPointVector() {
-		return Vector2(this->length / 2.0, height / 2.0);
+		return Vector2(this->length / 2.0f, height / 2.0f);
 	}
 
 	bool CollisionRectangle::isCollidingWith(Collider *otherCollider)
